Move common utility functions and constants to a `utils` folder
