+++ packages/components/src/calendar/index.tsx
@@ -1,9 +1,10 @@
 import { useMemo, useState } from 'react';
 import { DayPicker, TZDate } from 'react-day-picker';
 import { enUS } from 'react-day-picker/locale';
-import { Day } from './day-cell';
-import { useLocalizationProps } from './localization';
-import { useControlledValue } from './utils';
+import { COMMON_PROPS, MODIFIER_CLASSNAMES } from './utils/constants';
+import { clampNumberOfMonths } from './utils/misc';
+import { useControlledValue } from './utils/use-controlled-value';
+import { useLocalizationProps } from './utils/use-localization-props';
 import type {
 	DateCalendarProps,
 	DateRangeCalendarProps,
@@ -11,69 +12,6 @@ import type {
 	OnSelectHandler,
 } from './types';
 import './styles.scss';
-
-const BASE_CLASSNAME = 'a8c-components-calendar';
-const CLASSNAMES = {
-	root: BASE_CLASSNAME,
-	day: `${ BASE_CLASSNAME }__day`,
-	day_button: `${ BASE_CLASSNAME }__day-button`,
-	caption_label: `${ BASE_CLASSNAME }__caption-label`,
-	button_next: `${ BASE_CLASSNAME }__button-next`,
-	button_previous: `${ BASE_CLASSNAME }__button-previous`,
-	chevron: `${ BASE_CLASSNAME }__chevron`,
-	nav: `${ BASE_CLASSNAME }__nav`,
-	month_caption: `${ BASE_CLASSNAME }__month-caption`,
-	months: `${ BASE_CLASSNAME }__months`,
-	month_grid: `${ BASE_CLASSNAME }__month-grid`,
-	weekday: `${ BASE_CLASSNAME }__weekday`,
-	today: `${ BASE_CLASSNAME }__day--today`,
-	selected: `${ BASE_CLASSNAME }__day--selected`,
-	disabled: `${ BASE_CLASSNAME }__day--disabled`,
-	hidden: `${ BASE_CLASSNAME }__day--hidden`,
-	range_start: `${ BASE_CLASSNAME }__range-start`,
-	range_end: `${ BASE_CLASSNAME }__range-end`,
-	range_middle: `${ BASE_CLASSNAME }__range-middle`,
-	weeks_before_enter: `${ BASE_CLASSNAME }__weeks-before-enter`,
-	weeks_before_exit: `${ BASE_CLASSNAME }__weeks-before-exit`,
-	weeks_after_enter: `${ BASE_CLASSNAME }__weeks-after-enter`,
-	weeks_after_exit: `${ BASE_CLASSNAME }__weeks-after-exit`,
-	caption_after_enter: `${ BASE_CLASSNAME }__caption-after-enter`,
-	caption_after_exit: `${ BASE_CLASSNAME }__caption-after-exit`,
-	caption_before_enter: `${ BASE_CLASSNAME }__caption-before-enter`,
-	caption_before_exit: `${ BASE_CLASSNAME }__caption-before-exit`,
-};
-const MODIFIER_CLASSNAMES = {
-	preview: `${ BASE_CLASSNAME }__day--preview`,
-	preview_start: `${ BASE_CLASSNAME }__day--preview-start`,
-	preview_end: `${ BASE_CLASSNAME }__day--preview-end`,
-};
-
-const COMMON_PROPS = {
-	animate: true,
-	// Only show days in the current month
-	showOutsideDays: false,
-	// Hide week number column
-	showWeekNumber: false,
-	// Show weekdays row
-	hideWeekdays: false,
-	// Month and year caption are not interactive
-	captionLayout: 'label',
-	// Show a variable number of weeks depending on the month
-	fixedWeeks: false,
-	// Show navigation buttons
-	hideNavigation: false,
-	// Class names
-	classNames: CLASSNAMES,
-	// Default role
-	role: 'application',
-	components: {
-		Day,
-	},
-} as const;
-
-function clampNumberOfMonths( numberOfMonths: number ) {
-	return Math.min( 3, Math.max( 1, numberOfMonths ) );
-}
 
 export const DateCalendar = ( {
 	defaultSelected,
