+++ packages/components/src/calendar/utils/misc.ts
@@ -1,0 +1,3 @@
+export function clampNumberOfMonths( numberOfMonths: number ) {
+	return Math.min( 3, Math.max( 1, numberOfMonths ) );
+}
